app/requirements.txt
Django==5.1.5
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
django-environ==0.11.2
Pillow==11.1.0
pytest==8.3.4
pytest-django==4.9.0
flake8==7.1.1
